requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
discord.py==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
websockets==11.0.3
yarl==1.9.2
